--- SpringBoot/src/main/java/Baemin/News_Deliver/Global/Batch/BatchConfig.java
@@ -1,4 +1,0 @@
-package Baemin.News_Deliver.Global.Batch;
-
-public class BatchConfig {
-}
