+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/Batch/reader.java
@@ -1,0 +1,4 @@
+package Baemin.News_Deliver.Global.News.Batch;
+
+public class reader {
+}
