+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/ElasticSearch/dto/ElasticSearchReader.java
@@ -1,0 +1,4 @@
+package Baemin.News_Deliver.Global.News.ElasticSearch.dto;
+
+public class ElasticSearchReader {
+}
