+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/ElasticSearch/entity/ElasticsearchDocument.java
@@ -1,0 +1,4 @@
+package Baemin.News_Deliver.Global.News.ElasticSearch.entity;
+
+public class ElasticsearchDocument {
+}
