+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/ElasticSearch/repository/ElasticSearchWriter.java
@@ -1,0 +1,4 @@
+package Baemin.News_Deliver.Global.News.ElasticSearch.repository;
+
+public interface ElasticSearchWriter {
+}
