+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/JPAINSERT/BatchCall.java
@@ -1,5 +1,6 @@
-package Baemin.News_Deliver.Global.Batch.service;
+package Baemin.News_Deliver.Global.News.JPAINSERT;
 
+import Baemin.News_Deliver.Global.News.JPAINSERT.service.BatchService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
