+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/JPAINSERT/dto/NewsResponseDTO.java
@@ -1,4 +1,4 @@
-package Baemin.News_Deliver.Global.Batch.dto;
+package Baemin.News_Deliver.Global.News.JPAINSERT.dto;
 
 import lombok.*;
