+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/News/JPAINSERT/repository/NewsRepository.java
@@ -1,6 +1,6 @@
-package Baemin.News_Deliver.Global.Batch.repository;
+package Baemin.News_Deliver.Global.News.JPAINSERT.repository;
 
-import Baemin.News_Deliver.Global.Batch.entity.News;
+import Baemin.News_Deliver.Global.News.JPAINSERT.entity.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsRepository extends JpaRepository<News, Long> {
