+++ SpringBoot/src/main/java/Baemin/News_Deliver/Global/Scheduler/SchedulerConfig.java
@@ -1,5 +1,6 @@
 package Baemin.News_Deliver.Global.Scheduler;
 
+import Baemin.News_Deliver.Global.News.JPAINSERT.service.BatchService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -12,15 +13,15 @@ import java.time.LocalDateTime;
 @Slf4j
 public class SchedulerConfig {
 
+    private final BatchService batchService;
+
     //@Scheduled(cron = "0 0 0 * * *") // 매일 자정마다 로드
-    //@Scheduled(cron = "*/3 * * * * *") //[테스트용] 스케쥴러 동작 확인용 3초마다 테스트
+    @Scheduled(cron = "* */15 * * * *") //[테스트용] 스케쥴러 동작 확인용 3초마다 테스트
     public void dataBatchRun() {
         log.info("[Scheduler] database 배치 실행됨 - 시간: {}", LocalDateTime.now());
 
         //배치 코드 넘기기
-
-
-
+        batchService.batch();
     }
 
 }
